refactor: arranged common components
