+++ frontend/components/uploader/sidebar-vae/upload-file/upload-file.tsx
@@ -4,8 +4,55 @@ import { Form } from "react-bootstrap";
 import { useSelector } from "react-redux";
 import { useDispatch } from "react-redux";
 import { RootState } from "../../redux/store";
-import parseSelex from "~/components/common/parse-selex";
 import { apiClient } from "~/services/api-client";
+
+type ParseResult =
+  | {
+      status: "success";
+      data: string[];
+    }
+  | {
+      status: "error";
+      message: string;
+    };
+
+const parseSelex = async (file: File) => {
+  const fileType = file.name.split(".").pop() as string;
+  const fileData = await file.text();
+
+  if (fileType !== "fasta" && fileType !== "fastq") {
+    return {
+      status: "error",
+      message: "File type not supported",
+    } as ParseResult;
+  }
+
+  let regex;
+  let match: RegExpExecArray | null;
+  if (fileType === "fasta") {
+    regex = /^>[^\n\r]+[\n\r]+([ACGTUacgtu\n\r]+)$/gm;
+  } else {
+    regex = /^@[^\n\r]+[\n\r]+([ACGTUacgtu\n\r]+)$/gm;
+  }
+
+  let seqs: string[] = [];
+  while ((match = regex.exec(fileData))) {
+    const seq = match[1].replace(/\n/g, "");
+    seqs.push(seq);
+  }
+
+  if (seqs.length === 0) {
+    return {
+      status: "error",
+      message: "No sequences found",
+    } as ParseResult;
+  }
+
+  return {
+    status: "success",
+    data: seqs,
+  } as ParseResult;
+};
 
 type Props = {
   setVaeFile: React.Dispatch<React.SetStateAction<File | null>>;
